\name{create_collection}
\alias{create_collection}
\title{Creates a collection from a list of total impact ids (not working)}
\usage{
  create_collection(title = NULL, objects = NULL)
}
\arguments{
  \item{title}{Title of the collection}

  \item{objects}{a \code{list} of total impact ids.}
}
\value{
  collection id
}
\description{
  Creates a collection from a list of total impact ids (not
  working)
}
\examples{
\dontrun{
create_collection(title="my collection", list_of_items) # list must be valid total impact IDs.
}
}

