\name{github_heatmap}
\alias{github_heatmap}
\title{Generates a heatmap plot visualizing GitHub repos.}
\usage{
  github_heatmap(df)
}
\arguments{
  \item{df}{Input \code{data.frame}. Must be generated by
  \link{github_report}}
}
\value{
  \code{ggplot}
}
\description{
  Takes output of github_reports and makes a somewhat nice
  visualization
}
\examples{
\dontrun{
df <- github_report("d4npn7")
# To visualize this, then use
github_heatmap(df)
}
}

