\name{github_plot}
\alias{github_plot}
\title{Generates a plot visualizing GitHub repos.}
\usage{
  github_plot(df)
}
\arguments{
  \item{df}{Input \code{data.frame}. Must be generated by
  \link{github_report}}
}
\value{
  \code{ggplot}
}
\description{
  Takes output of github_reports and makes a nice
  visualization
}
\examples{
\dontrun{
df <- github_report("d4npn7")
# To visualize this, then use
github_plot(df)
}
}

