\name{metrics}
\alias{metrics}
\title{Returns the metrics for a valid total impact ID}
\usage{
  metrics(id = NA)
}
\arguments{
  \item{id}{= NA A total impact ID}
}
\description{
  Returns the metrics for a valid total impact ID
}
\examples{
\dontrun{
my_id <- tiid('10.1890/ES11-00339.1')
metrics(my_id)
}
}
\author{
  Karthik Ram \email{karthik.ram@gmail.com}
}
\seealso{
  \code{tiid}
}

