\name{metrics}
\alias{metrics}
\title{Returns the metrics for a valid Impact Story ID}
\usage{
  metrics(id = NA)
}
\arguments{
  \item{id}{= NA A Impact Story ID}
}
\description{
  Returns the metrics for a valid Impact Story ID
}
\examples{
\dontrun{
my_id <- ISid('10.1890/ES11-00339.1')
metrics(my_id)
}
}
\author{
  Karthik Ram \email{karthik.ram@gmail.com}
}
\seealso{
  \code{ISid}
}

