\name{metrics}
\alias{metrics}
\title{Returns the metrics for a valid total impact ID}
\usage{
  metrics(id = NA)
}
\arguments{
  \item{id}{= NA A total impact ID}
}
\description{
  Returns the metrics for a valid total impact ID
}
\examples{
\dontrun{
metrics(id = '0df8aa0eb2c911e19e181231381b0f5a')
}
}
\seealso{
  \code{tiid}
}

